SCPDriver: Make hashing routine user-definable
This introduces one new virtual method `getHashOf`,
which takes a vector of a vector of bytes - to be
able to hash a concatenation of parameters.

It returns a `Hash`, which is defined in `Stellar-types.x`.
Implementors of SCPDriver can define their own hash type
by redefining it in this XDR header file.

Co-authored-by: MonsieurNicolas <nicolas@stellar.org>

Resolves #2724